Migrate Mutation.schemaPublish
