+++ packages/services/api/src/modules/schema/resolvers.ts
@@ -32,7 +32,6 @@ import {
   Kind,
   print,
 } from 'graphql';
-import { parseResolveInfo } from 'graphql-parse-resolve-info';
 import { z } from 'zod';
 import { CriticalityLevel } from '@graphql-inspector/core';
 import type * as Types from '../../__generated__/types';
@@ -141,40 +140,6 @@ function __isTypeOf<
 
 export const resolvers: SchemaModule.Resolvers = {
   Mutation: {
-    async schemaPublish(_, { input }, { injector, request }, info) {
-      const [organization, project, target] = await Promise.all([
-        injector.get(OrganizationManager).getOrganizationIdByToken(),
-        injector.get(ProjectManager).getProjectIdByToken(),
-        injector.get(TargetManager).getTargetIdByToken(),
-      ]);
-
-      // We only want to resolve to SchemaPublishMissingUrlError if it is selected by the operation.
-      // NOTE: This should be removed once the usage of cli versions that don't request on 'SchemaPublishMissingUrlError' is becomes pretty low.
-      const parsedResolveInfoFragment = parseResolveInfo(info);
-      const isSchemaPublishMissingUrlErrorSelected =
-        !!parsedResolveInfoFragment?.fieldsByTypeName['SchemaPublishMissingUrlError'];
-
-      const result = await injector.get(SchemaPublisher).publish(
-        {
-          ...input,
-          service: input.service?.toLowerCase(),
-          organization,
-          project,
-          target,
-          isSchemaPublishMissingUrlErrorSelected,
-        },
-        request.signal,
-      );
-
-      if ('changes' in result) {
-        return {
-          ...result,
-          changes: result.changes,
-        };
-      }
-
-      return result;
-    },
     async schemaDelete(_, { input }, { injector, request }) {
       const [organization, project, target] = await Promise.all([
         injector.get(OrganizationManager).getOrganizationIdByToken(),
